backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
sqlalchemy
passlib[bcrypt]
python-jose
python-dotenv
